# docs/openapi/flowcept-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flowcept Webservice API",
    "description": "Read-only REST API for Flowcept provenance data. Provides workflows, tasks, and objects endpoints with query support.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/objects": {
      "get": {
        "tags": [
          "objects"
        ],
        "summary": "List Objects",
        "description": "List objects with optional basic filters.",
        "operationId": "list_objects_api_v1_objects_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "object_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Object Id"
            }
          },
          {
            "name": "workflow_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workflow Id"
            }
          },
          {
            "name": "task_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task Id"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "filter_json",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Filter Json"
            }
          },
          {
            "name": "include_data",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/objects/{object_id}": {
      "get": {
        "tags": [
          "objects"
        ],
        "summary": "Get Object",
        "description": "Get latest version of an object by id.",
        "operationId": "get_object_api_v1_objects__object_id__get",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            }
          },
          {
            "name": "include_data",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Object Api V1 Objects  Object Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/objects/{object_id}/versions/{version}": {
      "get": {
        "tags": [
          "objects"
        ],
        "summary": "Get Object Version",
        "description": "Get a specific object version by id and version number.",
        "operationId": "get_object_version_api_v1_objects__object_id__versions__version__get",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          },
          {
            "name": "include_data",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Object Version Api V1 Objects  Object Id  Versions  Version  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/objects/{object_id}/download": {
      "get": {
        "tags": [
          "objects"
        ],
        "summary": "Download Object",
        "description": "Download object payload as binary.",
        "operationId": "download_object_api_v1_objects__object_id__download_get",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/objects/{object_id}/versions/{version}/download": {
      "get": {
        "tags": [
          "objects"
        ],
        "summary": "Download Object Version",
        "description": "Download a specific object payload version as binary.",
        "operationId": "download_object_version_api_v1_objects__object_id__versions__version__download_get",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/objects/{object_id}/history": {
      "get": {
        "tags": [
          "objects"
        ],
        "summary": "Get Object History",
        "description": "Get object metadata history (latest-first).",
        "operationId": "get_object_history_api_v1_objects__object_id__history_get",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/objects/query": {
      "post": {
        "tags": [
          "objects"
        ],
        "summary": "Query Objects",
        "description": "Run an advanced read-only object query.",
        "operationId": "query_objects_api_v1_objects_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/query/{scope}": {
      "post": {
        "tags": [
          "query"
        ],
        "summary": "Query Scope",
        "description": "Run a read-only advanced query over a constrained collection scope.",
        "operationId": "query_scope_api_v1_query__scope__post",
        "parameters": [
          {
            "name": "scope",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "workflows",
                "tasks",
                "objects",
                "models",
                "datasets"
              ],
              "type": "string",
              "title": "Scope"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectQueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AggregationSpec": {
        "properties": {
          "operator": {
            "type": "string",
            "enum": [
              "avg",
              "sum",
              "min",
              "max"
            ],
            "title": "Operator"
          },
          "field": {
            "type": "string",
            "minLength": 1,
            "title": "Field"
          }
        },
        "type": "object",
        "required": [
          "operator",
          "field"
        ],
        "title": "AggregationSpec",
        "description": "Aggregation operator and source field."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ListResponse": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Items"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "count",
          "limit"
        ],
        "title": "ListResponse",
        "description": "Generic list envelope for collection endpoints."
      },
...
      "ObjectQueryRequest": {
        "properties": {
          "filter": {
            "additionalProperties": true,
            "type": "object",
            "title": "Filter"
          },
          "projection": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projection"
          },
          "limit": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Limit",
            "default": 100
          },
          "sort": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SortSpec"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort"
          },
          "aggregation": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AggregationSpec"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aggregation"
          },
          "remove_json_unserializables": {
            "type": "boolean",
            "title": "Remove Json Unserializables",
            "default": true
          },
          "include_data": {
            "type": "boolean",
            "title": "Include Data",
            "default": false
          }
        },
        "type": "object",
        "title": "ObjectQueryRequest",
        "description": "Object query payload with optional payload inclusion."
      },
...
      "SortSpec": {
        "properties": {
          "field": {
            "type": "string",
            "minLength": 1,
            "title": "Field"
          },
          "order": {
            "type": "integer",
            "enum": [
              1,
              -1
            ],
            "title": "Order",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "field"
        ],
        "title": "SortSpec",
        "description": "Sort field/order pair."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}